// apps/io-func-sign-user/src/infra/http/models/CreateFilledDocumentBody.ts
/**
 * Do not edit this file it is auto-generated by io-utils / gen-api-models.
 * See https://github.com/pagopa/io-utils
 */
/* eslint-disable  */

import { Email } from "./Email";
import * as t from "io-ts";
import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";

// required attributes
const CreateFilledDocumentBodyR = t.interface({
  document_url: NonEmptyString,

  email: Email,

  name: NonEmptyString,

  family_name: NonEmptyString
});

// optional attributes
const CreateFilledDocumentBodyO = t.partial({});

export const CreateFilledDocumentBody = t.exact(
  t.intersection(
    [CreateFilledDocumentBodyR, CreateFilledDocumentBodyO],
    "CreateFilledDocumentBody"
  )
);

export type CreateFilledDocumentBody = t.TypeOf<
  typeof CreateFilledDocumentBody
>;